Use functional programming instead of interfaces for managing callbacks from various forms.
The setup position screen now executes a callback passed in by the client when the position is submitted.  This allows for clients to execute their own behavior in response to the position set up on the cube.
Also moved RunnerScreen to the main Cubing project so any client can easily reuse it.